requirements.txt
appdirs==1.4.3
appnope==0.1.0
APScheduler==3.3.1
backports-abc==0.5
backports.shutil-get-terminal-size==1.0.0
BeautifulSoup==3.2.1
beautifulsoup4==4.6.0
bleach==2.0.0
bs4==0.0.1
certifi==2017.4.17
chardet==3.0.3
configparser==3.5.0
cycler==0.10.0
decorator==4.0.11
Django==1.10.7
django-contrib-comments==1.8.0
django-filter==1.0.4
django-modeltranslation==0.12.1
djangorestframework==3.6.3
ecdsa==0.13
entrypoints==0.2.2
enum34==1.1.6
Fabric==1.10.2
feedparser==5.2.1
filebrowser-safe==0.4.7
funcsigs==1.0.2
functools32==3.2.3.post2
future==0.16.0
futures==3.1.1
glog==0.3.1
grappelli-safe==0.4.6
html5lib==0.999999999
idna==2.5
ipykernel==4.6.1
ipython==5.0.0
ipython-genutils==0.2.0
ipywidgets==6.0.0
Jinja2==2.9.6
joblib==0.11
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.0.1
jupyter-console==5.1.0
jupyter-core==4.3.0
Markdown==2.6.8
MarkupSafe==1.0
matplotlib==2.0.2
Mezzanine==4.2.3
mezzanine-mdown==0.1a3
mezzanine-pagedown==1.0
mistune==0.7.4
mysqlclient==1.3.7
nbconvert==5.2.1
nbformat==4.3.0
nose==1.3.7
notebook==5.0.0
numpy==1.12.1
oauthlib==2.0.2
olefile==0.44
packaging==16.8
pandas==0.20.1
pandocfilters==1.4.1
paramiko==1.16.0
pathlib2==2.2.1
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.1.1
prompt-toolkit==1.0.14
ptyprocess==0.5.1
pycrypto==2.6.1
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.6.0
python-gflags==3.1.1
pytz==2017.2
pyzmq==16.0.2
qtconsole==4.3.0
requests==2.16.3
requests-oauthlib==0.8.0
scandir==1.5
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.10.0
solid==0.2.0
subprocess32==3.2.7
terminado==0.6
testpath==0.3.1
tornado==4.5.1
traitlets==4.3.2
tzlocal==1.4
urllib3==1.21.1
virtualenv==15.1.0
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==2.0.0
